Fix crashloop when using touchscreen
The touchscreen component needs to access the display's width and height.
The code was asking the tft_espi library for those, but at the time the
touchscreen needs the dimensions, the tft_espi has not yet been initialized;
and this caused a crash at startup.

Using the configuration variables ot set some defaults to be used when the
library has not yet been instantiated.