Fix compilation for esphome >= 2023.12.0
In the last release of ESPHome, some small changes were done to the display
component, to make subcomponents leaner. With those changes, compilation breaks
if a display is trying to register as a component itself (this is now done
automatically when registering the display).

Also, DisplayBuffer now automatically inherits from PollingComponent, so there is no need to explicitly inherit from that.